+++ lib/whois/record/parser/whois.bn.rb
@@ -1,0 +1,70 @@
+#--
+# Ruby Whois
+#
+# An intelligent pure Ruby WHOIS client and parser.
+#
+# Copyright (c) 2009-2013 Simone Carletti <weppos@weppos.net>
+#++
+
+
+require 'whois/record/parser/base'
+
+
+module Whois
+  class Record
+    class Parser
+
+      # Parser for the whois.bn server.
+      #
+      # @note This parser is just a stub and provides only a few basic methods
+      #   to check for domain availability and get domain status.
+      #   Please consider to contribute implementing missing methods.
+      #
+      # @see Whois::Record::Parser::Example
+      #   The Example parser for the list of all available methods.
+      #
+      class WhoisBn < Base
+
+        property_supported :status do
+          if available?
+            :available
+          else
+            :registered
+          end
+        end
+
+        property_supported :available? do
+          !!(content_for_scanner =~ /^No records matching .+ found/)
+        end
+
+        property_supported :registered? do
+          !available?
+        end
+
+
+        property_supported :created_on do
+          if content_for_scanner =~ /Created:\s+(.+)\n/
+            Time.parse($1)
+          end
+        end
+
+        property_supported :updated_on do
+          if content_for_scanner =~ /Last Updated:\s+(.+)\n/
+            Time.parse($1)
+          end
+        end
+
+        property_not_supported :expires_on
+
+
+        property_supported :nameservers do
+          content_for_scanner.scan(/Name Server \d:\s+(.+)\n/).flatten.map do |name|
+            Record::Nameserver.new(:name => name)
+          end
+        end
+
+      end
+
+    end
+  end
+end